Added the ability to "reuse" unstarted builds
Multiple consecutive builds will always trigger a build that has to
be run by a worker. Projects with long running test suites and/or
many configured matrix builds will cause a serious slow-down if many
builds are triggered at once (i.e. through multiple pushes).
This patch allows Travis to reuse a build of a project if it hasn't been
started yet. This will cause only the latest of the pushed commits to be
checked out and run.