This adds a welcome message for new users
I was lost as to what to do with a project after I had signed up.
Hopefully this will help people that come after me.

The template lives in templates/tools because the pages break if it lives in templates/ and there wasn't a better place for it.

It also breaks stuff for the github user called "welcome", but they have no repos so i don't care.