Adding mock for filter log.
For integration tests, for instance, it's more important to verify the workflow, but not internal actions of complex functions, especially ones like filterLog. We mock it in order to just return what it receives and do zero processing because otherwise tests will become extremely complicated to write read and modify.