Adding path elements to controller. We should actually always to be able to retrieve those. Maybe we should pass every single one of them to model or underlying views. That may require some thinking.
Other possibility is to get the "build" logic out of the builds themselves to the controller level (wether we're rendering log or no, wether we're rendering matrix build or something else etc). That way we will be able to control these things on the same level without building dependency chains.