Work around repository/build chicken-and-egg.
A repository has a computed property `lastBuild` which is based on a lastBuildId (which is passed along with an event). A build has a computed property `repository` which is based on a repositoryId (which is also passed with the same event).

When a repository is updated with the attributes from the event data first it will try to update the lastBuild property and find the respective build. As it might not (yet) be present locally it will reach out to the server.

The same happens when the build is updated with the passed attributes first. It will try to load it's repository and thus potentially reach out to the server.

No idea how to solve this problem in sproutcore. I would guess there's a way to temporarily deactivate the runloop or similar so computed properties would only get updated after we've created both records?