Add proper redirect handling for sign in
On pages that require authentication (like the profile page) we currently
still simply redirect to the root url without any notice.

This change makes it so that when the user tries to access an authenticated
page they will be redirected to /sessions/new where a simple sign in page
is displayed.

After signing in via GitHub they will be then redirected to whatever url
they originally wanted to access.

![](https://img.skitch.com/20120825-pn3b49xisb3m3g8a7dbq9kydgm.jpg)